requirements.txt
django==4.1.4
djangorestframework==3.14.0
django-cors-headers==3.13.0
whitenoise==6.2.0